Improvements
A posterior predictive sampler (which uses the parameters sample from the simulation) and a predictive graph generator tha compare both predictive sample and the data used in the model, was added to both models. Additionally, there was an improvement in graphic quality, mainly regarding font size, color adjustments, and the addition of labels.